pyproject.toml
[project]
name = "openstack-mcp-server"
version = "0.1.0"
description = "MCP server for OpenStack cloud infrastructure management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.2",
    "openstacksdk>=4.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openstack_mcp_server"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "ASYNC"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"